requirements.txt
fastapi
uvicorn[standard]
openai
reportlab
openpyxl
python-dotenv
google-generativeai
pydantic
